refactor: 💡 Move waitFor* to separate file
