+++ src/lib/wait.ts
@@ -1,0 +1,4 @@
+export const waitFor = (ms: number) =>
+  new Promise((resolve) => setTimeout(resolve, ms));
+
+export const waitForFrame = () => new Promise(requestAnimationFrame);
